refactor: move useOnClickOutside hook to hooks directory
